requirements/dev.txt
-r common.txt

# task automation
Fabric==1.10.1

# test libraries
coverage==3.7.1
mock==1.0.1

# celery
django-celery==3.1.16

# docs
sphinx_rtd_theme==0.1.6
sphinxcontrib-napoleon==0.3.1
